+++ lib/zip_mkstempm.c
@@ -1,6 +1,6 @@
 /*
-  zip_random_unix.c -- fill the user's buffer with random stuff (Unix version)
-  Copyright (C) 2016-2018 Dieter Baron and Thomas Klausner
+  zip_mkstempm.c -- mkstemp replacement that accepts a mode argument
+  Copyright (C) 2019 Dieter Baron and Thomas Klausner
 
   This file is part of libzip, a library to manipulate ZIP archives.
   The authors can be contacted at <libzip@nih.at>
@@ -31,75 +31,64 @@
   IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
+#include <sys/stat.h>
+#include <errno.h>
+#include <fcntl.h>
+#include <stdlib.h>
+#include <string.h>
+
 #include "zipint.h"
 
-#ifdef HAVE_CRYPTO
-#include "zip_crypto.h"
-#endif
-
-
-#ifdef HAVE_ARC4RANDOM
-
-#include <stdlib.h>
-
-#ifndef HAVE_SECURE_RANDOM
-ZIP_EXTERN bool
-zip_secure_random(zip_uint8_t *buffer, zip_uint16_t length) {
-    arc4random_buf(buffer, length);
-    return true;
-}
-#endif
-
-#ifndef HAVE_RANDOM_UINT32
-zip_uint32_t
-zip_random_uint32(void) {
-    return arc4random();
-}
-#endif
-
-#else /* HAVE_ARC4RANDOM */
-
-#ifndef HAVE_SECURE_RANDOM
-#include <fcntl.h>
-#include <unistd.h>
-
-ZIP_EXTERN bool
-zip_secure_random(zip_uint8_t *buffer, zip_uint16_t length) {
+/*
+ * create temporary file with same permissions as previous one;
+ * or default permissions if there is no previous file
+ */
+int
+_zip_mkstempm(char *path, int mode) {
     int fd;
+    char *start, *end, *xs;
 
-    if ((fd = open("/dev/urandom", O_RDONLY)) < 0) {
-	return false;
+    int xcnt = 0;
+
+    end = path + strlen(path);
+    start = end - 1;
+    while (start >= path && *start == 'X') {
+	xcnt++;
+	start--;
     }
 
-    if (read(fd, buffer, length) != length) {
-	close(fd);
-	return false;
+    if (xcnt == 0) {
+	errno = EINVAL;
+	return -1;
     }
 
-    close(fd);
-    return true;
+    start++;
+
+    for (;;) {
+        zip_uint32_t value = zip_random_uint32();
+
+	xs = start;
+
+	while (xs < end) {
+            char digit = value % 36;
+            if (digit < 10) {
+                *(xs++) = digit + '0';
+            }
+            else {
+                *(xs++) = digit - 10 + 'a';
+            }
+	    value /= 36;
+	}
+
+	if ((fd = open(path, O_CREAT | O_EXCL | O_RDWR | O_CLOEXEC, mode == -1 ? 0666 : (mode_t)mode)) >= 0) {
+	    if (mode != -1) {
+		/* open() honors umask(), which we don't want in this case */
+		(void)chmod(path, (mode_t)mode);
+	    }
+	    return fd;
+	}
+	if (errno != EEXIST) {
+	    return -1;
+	}
+    }
 }
-#endif
-
-#ifndef HAVE_RANDOM_UINT32
-#include <stdlib.h>
-
-zip_uint32_t
-zip_random_uint32(void) {
-    static bool seeded = false;
-    
-    zip_uint32_t value;
-    
-    if (zip_secure_random((zip_uint8_t *)&value, sizeof(value))) {
-        return value;
-    }
-    
-    if (!seeded) {
-        srandom((unsigned int)time(NULL));
-    }
-    
-    return (zip_uint32_t)random();
-}
-#endif
-
-#endif /* HAVE_ARC4RANDOM */