+++ regress/nonrandomopen.c
@@ -1,5 +1,5 @@
 /*
-  nonrandomopen.c -- override zip_random
+  nonrandomopen.c -- override zip_secure_random
 
   Copyright (C) 2017-2018 Dieter Baron and Thomas Klausner
 
@@ -43,7 +43,7 @@
 #include "zipint.h"
 
 bool
-zip_random(zip_uint8_t *buffer, zip_uint16_t length) {
+zip_secure_random(zip_uint8_t *buffer, zip_uint16_t length) {
     memset(buffer, 0, length);
 
     return true;